refactor: TODO 주석 제거
- 자잘한 문법 경고 해결
