+++ tomcat/src/main/java/org/apache/coyote/http11/controller/AbstractController.java
@@ -3,6 +3,7 @@ package org.apache.coyote.http11.controller;
 import org.apache.coyote.http11.request.HttpRequest;
 import org.apache.coyote.http11.response.HttpResponse;
 import org.apache.coyote.http11.response.HttpResponseHeader;
+import org.apache.coyote.http11.response.HttpResponseStatus;
 
 import java.io.IOException;
 import java.net.URISyntaxException;
@@ -15,14 +16,18 @@ import java.util.Objects;
 public abstract class AbstractController implements Controller {
     @Override
     public HttpResponse service(HttpRequest request) throws Exception {
-        if (request.isGET()) {
-            return doGet(request);
+        try {
+            if (request.isGET()) {
+                return doGet(request);
 
+            }
+            if (request.isPOST()) {
+                return doPost(request);
+            }
+            return null;
+        } catch (Exception e) {
+            return handle500();
         }
-        if (request.isPOST()) {
-            return doPost(request);
-        }
-        return null; // TODO: 2023/09/08 고치기
     }
 
     protected HttpResponse doPost(HttpRequest request) throws Exception {
@@ -47,5 +52,12 @@ public abstract class AbstractController implements Controller {
         return HttpResponseHeader.TEXT_HTML_CHARSET_UTF_8;
     }
 
+    private HttpResponse handle500() throws IOException, URISyntaxException {
+        String responseBody = getHtmlFile(getClass().getResource("/static/500.html"));
+        HttpResponseHeader responseHeader = new HttpResponseHeader(
+                HttpResponseHeader.TEXT_HTML_CHARSET_UTF_8,
+                String.valueOf(responseBody.getBytes().length), null, null);
+        return HttpResponse.of(HttpResponseStatus.INTERNAL_SERVER_ERROR, responseHeader, responseBody);
+    }
 }
