+++ tomcat/src/main/java/org/apache/coyote/http11/request/HttpRequestHeader.java
@@ -3,7 +3,7 @@ package org.apache.coyote.http11.request;
 import java.util.Map;
 
 public class HttpRequestHeader {
-    final private Map<String, String> headers;
+    private final Map<String, String> headers;
 
     public HttpRequestHeader(final Map<String, String> headers) {
         this.headers = headers;
